Don't attempt OAuth2 if the server doesn't support it
A server that supports OAuth2 will respond to an unauthenticated request with an
HTTP 401 that includes a wwwauth header:

  WWW-Authenticate: Bearer error="invalid_request", error_description="No bearer token found in the request"

git-credential passes those through to helpers in the form of the wwwauth[]
array. Scan for that and look for one starting with "Bearer". If that's not
found, don't attempt to get an authorization token.