Update source installation instructions to use pipx instead of pip
pipx is a specialized installer for ad-hoc cli entrypoints. Both pipx and
pip call pull dependencies from PyPI. However, pipx will not let you
install libraries in a way that is visible to the system python
interpreter. Libraries needed by a cli entrypoint will be encapsulated in a
venv without impacting the system install of python.

Link: https://pipx.pypa.io/stable/comparisons/
Signed-off-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>